Moved require_spark_cli() functionality inside require_spark().
