+++ tests/utils.py
@@ -1,8 +1,13 @@
+import dcos.config
+import dcos.http
+import dcos.package
+
 import logging
 import os
 import re
 import shakedown
 import subprocess
+import urllib
 
 
 def _init_logging():
@@ -13,6 +18,9 @@ def _init_logging():
 
 _init_logging()
 LOGGER = logging.getLogger(__name__)
+DEFAULT_HDFS_TASK_COUNT=10
+HDFS_PACKAGE_NAME='beta-hdfs'
+HDFS_SERVICE_NAME='hdfs'
 SPARK_PACKAGE_NAME='spark'
 
 
@@ -22,6 +30,96 @@ def hdfs_enabled():
 
 def is_strict():
     return os.environ.get('SECURITY') == 'strict'
+
+
+def require_hdfs():
+    LOGGER.info("Ensuring HDFS is installed.")
+
+    _require_package(HDFS_PACKAGE_NAME, _get_hdfs_options())
+    _wait_for_hdfs()
+
+
+def require_spark():
+    LOGGER.info("Ensuring Spark is installed.")
+
+    _require_package(SPARK_PACKAGE_NAME, _get_spark_options())
+    _wait_for_spark()
+    _require_spark_cli()
+
+
+# This should be in shakedown (DCOS_OSS-679)
+def _require_package(pkg_name, options = {}):
+    pkg_manager = dcos.package.get_package_manager()
+    installed_pkgs = dcos.package.installed_packages(pkg_manager, None, None, False)
+
+    if any(pkg['name'] == pkg_name for pkg in installed_pkgs):
+        LOGGER.info("Package {} already installed.".format(pkg_name))
+    else:
+        LOGGER.info("Installing package {}".format(pkg_name))
+        shakedown.install_package(
+            pkg_name,
+            options_json=options,
+            wait_for_completion=True)
+
+
+def _wait_for_spark():
+    def pred():
+        dcos_url = dcos.config.get_config_val("core.dcos_url")
+        spark_url = urllib.parse.urljoin(dcos_url, "/service/spark")
+        status_code = dcos.http.get(spark_url).status_code
+        return status_code == 200
+
+    shakedown.wait_for(pred)
+
+
+def _require_spark_cli():
+    LOGGER.info("Ensuring Spark CLI is installed.")
+    installed_subcommands = dcos.package.installed_subcommands()
+    if any(sub.name == SPARK_PACKAGE_NAME for sub in installed_subcommands):
+        LOGGER.info("Spark CLI already installed.")
+    else:
+        LOGGER.info("Installing Spark CLI.")
+        shakedown.run_dcos_command('package install --cli {}'.format(
+            SPARK_PACKAGE_NAME))
+
+
+def _get_hdfs_options():
+    if is_strict():
+        options = {'service': {'principal': 'service-acct', 'secret_name': 'secret'}}
+    else:
+        options = {"service": {}}
+
+    options["service"]["beta-optin"] = True
+    return options
+
+
+def _wait_for_hdfs():
+    shakedown.wait_for(_is_hdfs_ready, ignore_exceptions=False, timeout_seconds=900)
+
+
+def _is_hdfs_ready(expected_tasks = DEFAULT_HDFS_TASK_COUNT):
+    running_tasks = [t for t in shakedown.get_service_tasks(HDFS_SERVICE_NAME) \
+                     if t['state'] == 'TASK_RUNNING']
+    return len(running_tasks) >= expected_tasks
+
+
+def _get_spark_options():
+    if hdfs_enabled():
+        options = {"hdfs":
+                       {"config-url":
+                            "http://api.hdfs.marathon.l4lb.thisdcos.directory/v1/endpoints"}}
+    else:
+        options = {}
+
+    if is_strict():
+        options.update({'service':
+                            {"principal": "service-acct"},
+                        "security":
+                            {"mesos":
+                                 {"authentication":
+                                      {"secret_name": "secret"}}}})
+
+    return options
 
 
 def run_tests(app_url, app_args, expected_output, args=[]):
